Bug 2876 fix
git-svn-id: http://code.jaspersoft.com/svn/repos/jasperreports/trunk/jasperreports@2309 76e25fc1-081a-0410-bb1b-c96bc70638b7

+++ src/net/sf/jasperreports/engine/export/JRXlsExporter.java
@@ -1128,7 +1128,9 @@ public class JRXlsExporter extends JRXlsAbstractExporter
 				cell = row.createCell((short)colIndex);
 				cell.setCellStyle(cellStyle);
 
-				HSSFClientAnchor anchor = new HSSFClientAnchor(0, 0, 0, 0, (short) colIndex, (short) rowIndex, (short) (colIndex + gridCell.getColSpan()), (short) (rowIndex + gridCell.getRowSpan()));
+				HSSFClientAnchor anchor = new HSSFClientAnchor(0, 0, 0, 0, 
+						(short) colIndex, rowIndex, 
+						(short) (colIndex + gridCell.getColSpan()), rowIndex + gridCell.getRowSpan());
 				anchor.setAnchorType(2);
 				//pngEncoder.setImage(bi);
 				//int imgIndex = workbook.addPicture(pngEncoder.pngEncode(), HSSFWorkbook.PICTURE_TYPE_PNG);
